net: shell: Add route command
Move "route" command to a separate file.
